P2022-1077 [Tech] Package naming -> P2022-1192 Packages refactor
