+++ app/src/main/AndroidManifest.xml
@@ -12,14 +12,14 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.App.Starting">
         <activity
-            android:name=".offline.OfflineActivity"
+            android:name=".offline.presentation.view.OfflineActivity"
             android:configChanges="orientation|keyboardHidden|screenSize"
             android:exported="false"
             android:label="@string/title_activity_offline"
             android:theme="@style/Theme.AppCompat"
             />
         <activity
-            android:name=".MainActivity"
+            android:name="com.intive.patronage.retro.main.presentation.view.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.App.Starting">
             <intent-filter>
