+++ app/src/main/java/com/intive/patronage/retro/board/presentation/viewModel/BoardViewModel.kt
@@ -1,4 +1,4 @@
-package com.intive.patronage.retro.board_fragment
+package com.intive.patronage.retro.board.presentation.viewModel
 
 import androidx.lifecycle.ViewModel
