+++ app/src/main/java/com/intive/patronage/retro/common/dependencyInjection/Modules.kt
@@ -1,10 +1,11 @@
-package com.intive.patronage.retro
+package com.intive.patronage.retro.common.dependencyInjection
 
 import android.content.Context
 import android.net.ConnectivityManager
-import com.intive.patronage.retro.auth.Auth
-import com.intive.patronage.retro.auth.AuthRepository
-import com.intive.patronage.retro.util.CheckNetworkConnect
+import com.intive.patronage.retro.auth.model.repo.AuthRepository
+import com.intive.patronage.retro.auth.model.service.Auth
+import com.intive.patronage.retro.common.network.CheckNetworkConnect
+import com.intive.patronage.retro.main.presentation.viewModel.MainViewModel
 import org.koin.android.ext.koin.androidApplication
 import org.koin.android.ext.koin.androidContext
 import org.koin.androidx.viewmodel.dsl.viewModel
