+++ app/src/main/java/com/intive/patronage/retro/history/presentation/view/HistoryFragment.kt
@@ -1,4 +1,4 @@
-package com.intive.patronage.retro.board_fragment
+package com.intive.patronage.retro.history.presentation.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,20 +6,20 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.google.android.material.bottomappbar.BottomAppBar
-import com.intive.patronage.retro.MainActivity
 import com.intive.patronage.retro.R
-import com.intive.patronage.retro.databinding.BoardFragmentBinding
+import com.intive.patronage.retro.databinding.HistoryFragmentBinding
+import com.intive.patronage.retro.main.presentation.view.MainActivity
 
-class BoardsFragment : Fragment() {
+class HistoryFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ) = BoardFragmentBinding.inflate(inflater, container, false).root
+        savedInstanceState: Bundle?
+    ) = HistoryFragmentBinding.inflate(inflater, container, false).root
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         val bottomAppBar = (activity as MainActivity).findViewById<BottomAppBar>(R.id.bottomAppBar)
-        bottomAppBar.replaceMenu(R.menu.bottom_app_bar_menu_boards)
+        bottomAppBar.replaceMenu(R.menu.bottom_app_bar_menu_history)
     }
 }
