+++ app/src/main/java/com/intive/patronage/retro/history/presentation/viewModel/HistoryViewModel.kt
@@ -1,4 +1,4 @@
-package com.intive.patronage.retro.history_fragment
+package com.intive.patronage.retro.history.presentation.viewModel
 
 import androidx.lifecycle.ViewModel
