+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:id="@+id/drawerLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
+    tools:context=".main.presentation.view.MainActivity"
     android:fitsSystemWindows="true">
 
     <androidx.fragment.app.FragmentContainerView
