+++ app/src/main/res/layout/activity_offline.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".offline.OfflineActivity">
+    tools:context=".offline.presentation.view.OfflineActivity">
 
     <ImageView
         android:id="@+id/imageView"
