+++ app/src/main/res/layout/board_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".board_fragment.BoardsFragment">
+    tools:context=".board.presentation.view.BoardsFragment">
 
     <TextView
         android:layout_width="wrap_content"
