+++ app/src/main/res/layout/profile_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".profile_fragment.ProfileFragment">
+    tools:context=".profile.presentation.view.ProfileFragment">
 
     <TextView
         style="@style/demo_text"
